LayoutTests/platform/mac/fullscreen/video-controls-drag-expected.txt

This tests that the video element's controls are draggable in full screen. Press any key to continue.

EVENT(webkitfullscreenchange)
EXPECTED (endBox.left - startBox.left == '15'), OBSERVED '-385' FAIL
EXPECTED (endBox.top - startBox.top == '15'), OBSERVED '605' FAIL
END OF TEST
